Redirect Search to Product if Only One Result
For over ten years, the advanced search has not sent a search resulting in only one product to that product's actual page.  In https://www.zen-cart.com/showthread.php?86055-Search-Redirect-if-only-one-result-found&p=492888#post492888, a fix was proposed by DrByte that would allow the search to present the single item on its product page.  This fix works with 1.5.6c with minimal mods and demo items.  Example at netzencommerce.com.  Search for Russ or Responsive.